% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ALHM.R
\name{arid_model}
\alias{arid_model}
\title{Arid Model - Runs Spatially Distributed Hydrology Model}
\usage{
arid_model(
  date,
  ModelFolder,
  time_step = 0.5,
  simulation_length = NA,
  WatershedElements =
    "C:\\\\Thesis\\\\Arid-Land-Hydrology\\\\R\\\\Example\\\\MiniWatershedElements",
  mini = F,
  rainfall_created = F,
  save = T,
  All_save = T,
  crop = T,
  discharge = T
)
}
\arguments{
\item{date}{The date of rainfall event in form "YYYY-MM-DD"}

\item{ModelFolder}{Folder path to saved modeled outputs}

\item{time_step}{Time in minutes for algorithm step for model, will dynamically adjust}

\item{simulation_length}{Duration of model in minutes: NA = default entire event}

\item{WatershedElements}{Folder path where watershed specific elements are located.
DEM, Computational boundary, and rainfall storage}

\item{mini}{T/F Optional for miniature version of water}

\item{rainfall_created}{Optional if rainfall event has been created}

\item{save}{Optional if True will save plots in Model Folder}

\item{All_save}{Optional if TRUE will save important variables}

\item{crop}{Optional: if TRUE watershed boundary shapefile will clip watershed elements}

\item{discharge}{Optional: if TRUE script expects recorded discharge data in watershed elements for comparisons}
}
\value{
Creates stacked rasters of surface depth, velocity, and soil moisture
}
\description{
Arid Model - Runs Spatially Distributed Hydrology Model
}
\examples{
ModelFolder <- r"(C:\R\Example\SampleModel\Demo_Test)"
WatershedElements <- r"(C:\R\Example\DemoElements)"
# Do Not RUN
# arid_model(date, ModelFolder, time_step = 5,
#         WatershedElements = WatershedElements)

}
