% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/desertHydro-package.R
\docType{package}
\name{desertHydro-package}
\alias{desertHydro}
\alias{desertHydro-package}
\title{desertHydro: Spatially Distributed Hydrology Model for Arid Watersheds}
\description{
The desertHydro watershed model simulates the surface and shallow subsurface flow from an input rainfall event. The primary flow routing algorithm is modified from a Soil Moisture Routing (SMR) algorithm. The user inputs a digital elevation model (DEM), watershed boundary, (shapefile), soil characteristics, and precipitation data. The model outputs stacked tif files with each timestep as a layer including velocity, soil moisture, and surface depth. Optionally, these stacked rasters can be converted into gifs for visual aids. Note - This is still in early stages of development.
}
\author{
\strong{Maintainer}: Maxwell Miller \email{mmm683@nau.edu} (\href{https://orcid.org/0009-0000-2965-695X}{ORCID})

}
\keyword{internal}
