% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Discharge_Process.R
\name{discharge_rainfall_events}
\alias{discharge_rainfall_events}
\title{Discharge-Rainfall Comparisons
Determine the discharge amounts, lengths, and rainfall amounts for a set of given dates
Function is not very flexible in current state.}
\usage{
discharge_rainfall_events(
  discharge_file,
  ModelFolder,
  WatershedElements = model()@watershedPath,
  store = F
)
}
\arguments{
\item{discharge_file}{table containing a list of discharge dates}

\item{ModelFolder}{Path to save location}

\item{WatershedElements}{Path to location of saved stream data and rainfall data}

\item{store}{T/F If TRUE, saves elements to ModelFolder}
}
\value{
Returns and saves csv file into Model Folder. Optional saves recorded
discharge and rainfall events as excel spreadsheet.
}
\description{
Discharge-Rainfall Comparisons
Determine the discharge amounts, lengths, and rainfall amounts for a set of given dates
Function is not very flexible in current state.
}
\examples{
\dontrun{
# Change this to output folder
discharge_file <- "C:/path_to_discharge/discharge_file.csv"
ModelFolder <- r"(C:/Thesis/Arid-Land-Hydrology/R/Example/SampleModel)"
WatershedElements <- file.path("inst/extdata/DemoElements") # demo elements
discharge_events <- discharge_rainfall_events(discharge_file, ModelFolder,
                                      WatershedElements, store = T)
}
}
