% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{folderCheck}
\alias{folderCheck}
\title{Function to check and adjust the WatershedElements Folder
Wrapper function of system.file to check if the folder containing the model
elements exists and returns the path to the folder. Will default to package
elements if no file is found.}
\usage{
folderCheck(folder, example_file = "dem.tif", demo = T)
}
\arguments{
\item{folder}{Folder path to check}

\item{example_file}{Default "dem.tif'. File name + extension to check for within folder.}

\item{demo}{boolean. If TRUE, will use demo elements folder if input folder is not found}
}
\value{
Directory or folder path. Demo folder if demo = T.
}
\description{
Function to check and adjust the WatershedElements Folder
Wrapper function of system.file to check if the folder containing the model
elements exists and returns the path to the folder. Will default to package
elements if no file is found.
}
\examples{
\dontrun{
folderPath <- "./WatershedElements"
WatershedElements <- folderCheck(folderPath, "dem.tif", demo = T)
WatershedElements
}
# Returns demo folder
}
