% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{getCellCoords}
\alias{getCellCoords}
\title{Find the cell numbers in a raster from a given shapefile/vector.
This function is basically a wrapper for "cellFromXY" within the terra package}
\usage{
getCellCoords(vect, raster)
}
\arguments{
\item{vect}{shapefile}

\item{raster}{raster}
}
\value{
cell numbers for points within shapefile
}
\description{
Find the cell numbers in a raster from a given shapefile/vector.
This function is basically a wrapper for "cellFromXY" within the terra package
}
\examples{
/dontrun{
cell_numbers <- getCellCoords(vector, raster)
}

}
