% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/postAnalysis.R
\name{gifCreation}
\alias{gifCreation}
\title{Create GIFS}
\usage{
gifCreation(
  ModelFolder,
  rainfall_method = "",
  date = NULL,
  discharge = F,
  saveGraph = T,
  rainfall_only = F
)
}
\arguments{
\item{ModelFolder}{Location to save Model Folder}

\item{rainfall_method}{String: "guages","spatial", "synthetic", "goes based on
the input rainfall method}

\item{date}{String: Optional, will display the rainfall totals in GIF}

\item{discharge}{T/F: If T, will incorporate discharge data}

\item{saveGraph}{T/F: If T, will save graph to ModelFolder}

\item{rainfall_only}{T/F: If TRUE, will only create rainfall GIF}
}
\value{
Returns list containing surface and velocity GIFS.
}
\description{
Create GIFS
}
\examples{
\dontrun{
gifCreation(ModelFolder, saveGraph = T)
}
}
