% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/monsoon.R
\name{monsoon}
\alias{monsoon}
\title{Complete Model Runs}
\usage{
monsoon(
  ModelFolder = NULL,
  date = "2021-07-22",
  model_type = "default",
  rain_method = "gauges",
  depth_adjusted = "slope",
  surface_method = "soils+stream+geo",
  infiltration_method = "soils+geo+green",
  initial_soil_conditions = "normal",
  rain_adj = 1,
  surface_adj = 1,
  infiltration_adj = 1,
  courant = 0.8,
  time = NaN,
  overwrite = NA,
  velocity_method = "darcys",
  debug = F,
  ...
)
}
\arguments{
\item{ModelFolder}{Path to folder to put output files}

\item{date}{Date string of event in format "YYYY-MM-DD", default}

\item{model_type}{Sets infiltration and size parameters, default = "default"
Other options include:
"l-runoff" = large scale model with no infiltration,
"s-runoff" = small scale model with no infiltration,
"l-infiltration" = large scale model with infiltration,
"s-infiltration" = small scale model with infiltration}

\item{rain_method}{type of rainfall to be used. default = "gauges" from gauge
data found within the Watershed folder path.
Other options include:
"spatial" = spatially distributed rainfall based on gauge placements
"goes" = satellite derived rainfall estimates obtained from AWS servers for current
spatial extent of model}

\item{depth_adjusted}{Adjust the depth. Default "slope" method}

\item{surface_method}{Methods for adjusting the surface roughness. Defaults
"soils+stream+geo"}

\item{infiltration_method}{Methods for infiltration. Adjusts the infiltration
and chains with the method of infiltration. Often Green-Ampt method. Defaults
"soils+geo+green"}

\item{initial_soil_conditions}{Soil Saturation initially. Defaults "normal"}

\item{rain_adj}{\% change in rainfall. Default 1 (100\%) of recorded values}

\item{surface_adj}{\% change in surface roughness Default 1 (100\%) of recorded values}

\item{infiltration_adj}{\% change in infiltration Default 1 (100\%) of recorded values}

\item{courant}{the distance information travels within a mesh network. Values
should range between 0.1 and 1.0. The smaller the Courant number, the shorter
the time steps will be, which leads to more numerical stability. Default value is
0.8; however, research has shown that for arid regions Courant values as low as
0.1 are needed. The smaller the Courant number the longer simulations will take.}

\item{time}{Default NaN, which will calculate the length of the simulation based
upon the discharge and rainfall input data. Setting time (minutes) to an explicit value
will have the simulation run for only that amount of time. Helpful in testing
parameters without performing an entire rainfall simulation.}

\item{overwrite}{Default NA, will change depending on which model_type was selected}

\item{...}{Additional parameters (advanced)}
}
\value{
returns path to Model Folder
}
\description{
Function that allows multiple model runs and has the ability to work on
cluster nodes. Note: The functionality of the script is based upon original work
and is not flexible enough for everyday use.
}
\examples{
\dontrun{
# Change this to output folder
ModelFolder <- "Rainfall-2021-07-22"
model_output <- monsoon(ModelFolder = ModelFolder,
                        date = "2021-07-22",
                        model_type = "l-runoff",
                        rain_method ="gauges",
                        courant = 0.3)
}
}
