% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{polygonize}
\alias{polygonize}
\title{Wrapper function to dissolve and create vector of raster layer}
\usage{
polygonize(raster, path)
}
\arguments{
\item{raster}{Raster name, e.i. "dem.tif"}

\item{path}{file path to raster}
}
\value{
returns vector of polygon
}
\description{
Wrapper function to dissolve and create vector of raster layer
}
