% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Rainfall_Process.R
\name{rainfallTotalRain}
\alias{rainfallTotalRain}
\title{Gathers rainfall data based on date from spreadsheet}
\usage{
rainfallTotalRain(
  rainfall_folder,
  date,
  level = "day",
  total_col = "TOTAL",
  gauges = c("WATER-1", "WATER-2", "WATER-G"),
  write = F
)
}
\arguments{
\item{rainfall_folder}{Folder path of rainfall data}

\item{date}{Date string in format: "2021-07-13"}

\item{level}{Optional filter: can filter by "day", "hour", "minute" based on
utilizes lubridate package}

\item{total_col}{Name of total column: Default "TOTAL"}

\item{gauges}{Name of gauges spreadsheets:
gauges = c("WATER-1", "WATER-2", "WATER-G")}

\item{write}{T/F. If TRUE, will write output to rainfall_folder}
}
\value{
Returns data frame with rainfall data filtered by level
}
\description{
Gathers rainfall data based on date from spreadsheet
}
\examples{
rain_f <- dirname(system.file("extdata", "dem.tif", package = "desertHydro"))
rainfallTotalRain(rain_f, "2022-07-05", level = "day")
}
