% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{rasterCompile}
\alias{rasterCompile}
\title{Function that takes a bunch of rasters and combined them by name and time}
\usage{
rasterCompile(
  ModelFolder,
  layername,
  remove = T,
  time = T,
  overwrite = T,
  end = T
)
}
\arguments{
\item{ModelFolder}{path to folder containing files}

\item{layername}{string of layername that will combine files. e.g. "velocity"}

\item{remove}{T/F. If True will remove all of the separate files}

\item{time}{T/F. Default True will sort files in order of time.
Expects format "layername-time-time.tif"}
}
\value{
Returns combined raster stack of named files
}
\description{
Function that takes a bunch of rasters and combined them by name and time
}
\examples{
\dontrun{
# Format of function
rasterStack <- rasterCompile("Test-Folder", "velocity", remove = T)
}
}
