% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/createWatershedModel.R
\name{resizeShape}
\alias{resizeShape}
\title{Resizes and reprojects the shape of a spatial object based on a pair}
\usage{
resizeShape(
  spatialObject,
  extent_raster,
  boundary_path = NA,
  key = "MUSYM",
  save = FALSE,
  save_name = "",
  save_location = ""
)
}
\arguments{
\item{spatialObject}{SpatVector or SpatRaster to be resized and/or reprojected}

\item{extent_raster}{Spatial extent raster to reproject into}

\item{boundary_path}{Optional shapefile path location to crop raster to}

\item{key}{Key to land cover type}

\item{save}{T/F If T, saves adjust land cover raster}

\item{save_name}{If save = T, file name to save adjusted land cover raster -
"~/adjusted_land_cover.tif"}

\item{save_location}{If save = T, location to save adjusted land cover raster}
}
\value{
Returns a SpatRaster with reprojected and cropped to input extents
dimenstions
}
\description{
Resizes and reprojects the shape of a spatial object based on a pair
}
\examples{
\dontrun{
spatialObject <- terra::rast()
extent_raster <- terra::rast()
boundary_path <- terra::vect()
key <- "MUSYM"
new_land_cover <- resizeShape(spatialObject, extent_raster, boundary_path, key = key)
}
}
