% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/streamProcess.R
\name{smoothStream}
\alias{smoothStream}
\title{Function to take a collection of vector
streams lines and burn in the elevation profiles into the model}
\usage{
smoothStream(stream, demPath, outpath = NULL)
}
\arguments{
\item{stream}{SpatVector or file path containing the stream lines}

\item{demPath}{File path to SpatRaster}

\item{outpath}{Optional: Output path. Will be saved as "smooth_dem.tif"}
}
\value{
SpatRaster with stream network burned in.
}
\description{
Function to take a collection of vector
streams lines and burn in the elevation profiles into the model
}
\examples{
\dontrun{
smooth_dem <- smoothStream(streamVector, demPath)
}

}
