% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/streamProcess.R
\name{smoothVector}
\alias{smoothVector}
\title{Adjust stream channel elevations to remove pits}
\usage{
smoothVector(x, n = 1000, h_adj = 0.01)
}
\arguments{
\item{x}{vector of elevation values}

\item{n}{maximum number of iterations to fix elevations}

\item{h_adj}{additional increment to the smoothed locations Default
0.01.}
}
\value{
Smoothed vector of elevations
}
\description{
Adjust stream channel elevations to remove pits
}
\examples{
example_stream <- c(1,2,3,5,3,2,6)
smooth <- smoothVector(example_stream)

# If NA's are present
s2 <- c(1,2,4,5,NA,6)
smooth <- smoothVector(s2)
}
