% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/postAnalysis.R
\name{dischargeAnalysis}
\alias{dischargeAnalysis}
\title{Discharge Analysis after desertFlo simulation}
\usage{
dischargeAnalysis(
  ModelFolder,
  WatershedElements,
  time_step,
  simulation_length,
  discharge = T,
  store = T
)
}
\arguments{
\item{ModelFolder}{Model Folder with saved outputs}

\item{WatershedElements}{Watershed Folder with Watershed components. Assumes that}

\item{time_step}{Time in minutes}

\item{simulation_length}{Length of simulation in minutes}

\item{discharge}{T/F: Default TRUE: If TRUE, will look for observed discharge hydrographs}

\item{store}{T/F: Default TRUE: If TRUE, will store created plots in ModelFolder}
}
\value{
Returns discharge excel table and plots from model simulation
}
\description{
Discharge Analysis after desertFlo simulation
}
\examples{
See vignette
}
